# backend/ai-engine/openapi_schema.json
{"openapi": "3.1.0", "info": {"title": "Crime Intelligence AI Engine", "version": "0.1.0"}, "paths": {"/ai/ask": {"post": {"tags": ["ai-engine"], "summary": "Ask", "operationId": "ask_ai_ask_post", "parameters": [{"name": "authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AskRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StandardResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai/feedback": {"post": {"tags": ["ai-engine"], "summary": "Ai Feedback", "description": "10.8 Feedback Loop\nEndpoint to rate AI answers.", "operationId": "ai_feedback_ai_feedback_post", "parameters": [{"name": "query_id", "in": "query", "required": true, "schema": {"type": "string", "title": "Query Id"}}, {"name": "rating", "in": "query", "required": true, "schema": {"type": "integer", "title": "Rating"}}, {"name": "comments", "in": "query", "required": false, "schema": {"type": "string", "default": "", "title": "Comments"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai/transcribe": {"post": {"tags": ["Voice IO"], "summary": "Transcribe Audio", "operationId": "transcribe_audio_ai_transcribe_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_transcribe_audio_ai_transcribe_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai/speak": {"get": {"tags": ["Voice IO"], "summary": "Speak Text", "operationId": "speak_text_ai_speak_get", "parameters": [{"name": "text", "in": "query", "required": true, "schema": {"type": "string", "title": "Text"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai/history/{user_id}/{session_id}/export.pdf": {"get": {"tags": ["History"], "summary": "Export Conversation Pdf", "operationId": "export_conversation_pdf_ai_history__user_id___session_id__export_pdf_get", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}, {"name": "session_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Session Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai/documents/upload": {"post": {"tags": ["Documents"], "summary": "Upload Documents", "operationId": "upload_documents_ai_documents_upload_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_upload_documents_ai_documents_upload_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai/graph/visualize": {"get": {"tags": ["Graph Analysis"], "summary": "Visualize Graph", "description": "3.2 Link / Graph Visualization\nReturns nodes and edges for frontend network visualization (e.g., vis.js, D3).", "operationId": "visualize_graph_ai_graph_visualize_get", "parameters": [{"name": "fir_id", "in": "query", "required": false, "schema": {"type": "string", "description": "Filter graph by specific FIR ID", "title": "Fir Id"}, "description": "Filter graph by specific FIR ID"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai/graph/community": {"get": {"tags": ["Graph Analysis"], "summary": "Detect Communities", "description": "3.3 Community Detection\nUses simple traversal or graph algos to find clusters/syndicates of accused.", "operationId": "detect_communities_ai_graph_community_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/ai/graph/centrality": {"get": {"tags": ["Graph Analysis"], "summary": "Role Centrality", "description": "3.4 Role & Centrality Analysis\nIdentifies key suspects or central hubs in criminal networks based on degree centrality.", "operationId": "role_centrality_ai_graph_centrality_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/ai/graph/shortest-path": {"get": {"tags": ["Graph Analysis"], "summary": "Shortest Path", "description": "3.6 Social Network Analysis Queries\nFinds the shortest path between two entities to discover hidden links.", "operationId": "shortest_path_ai_graph_shortest_path_get", "parameters": [{"name": "source_id", "in": "query", "required": true, "schema": {"type": "string", "title": "Source Id"}}, {"name": "target_id", "in": "query", "required": true, "schema": {"type": "string", "title": "Target Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai/graph/geography": {"get": {"tags": ["Graph Analysis"], "summary": "Geographic Mapping", "description": "3.7 Geographic Relationship Mapping\nExtracts lat/lng from graph cases to map connected crimes geographically.", "operationId": "geographic_mapping_ai_graph_geography_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/ai/graph/temporal": {"get": {"tags": ["Graph Analysis"], "summary": "Temporal Dynamics", "description": "3.8 Temporal Dynamics\nAdds a time-window filter to view network evolution over time.", "operationId": "temporal_dynamics_ai_graph_temporal_get", "parameters": [{"name": "year", "in": "query", "required": false, "schema": {"type": "integer", "description": "Filter graph by specific year", "title": "Year"}, "description": "Filter graph by specific year"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai/graph/hidden-links": {"get": {"tags": ["Graph Analysis"], "summary": "Hidden Links", "description": "3.9 Hidden Link Discovery\nFinds non-obvious shared history through 2nd-degree traversal.", "operationId": "hidden_links_ai_graph_hidden_links_get", "parameters": [{"name": "suspect_id", "in": "query", "required": true, "schema": {"type": "string", "title": "Suspect Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai/graph/query": {"post": {"tags": ["Graph Analysis"], "summary": "Interactive Query", "description": "3.11 Interactive Graph Queries\nSafe parameterized endpoint for custom graph queries.", "operationId": "interactive_query_ai_graph_query_post", "parameters": [{"name": "cypher", "in": "query", "required": true, "schema": {"type": "string", "title": "Cypher"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai/graph/export": {"get": {"tags": ["Graph Analysis"], "summary": "Export Subgraph", "description": "3.12 Subgraph Export\nDownloads Neo4j subgraph as JSON.", "operationId": "export_subgraph_ai_graph_export_get", "parameters": [{"name": "fir_id", "in": "query", "required": true, "schema": {"type": "string", "title": "Fir Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai/profiling/story/{accused_id}": {"get": {"tags": ["Offender Profiling"], "summary": "Generate Offender Story", "description": "6.9 Offender 'Story' Generation (LLM Narrative Profile)\nGenerates a readable investigative brief of an offender using LLM.", "operationId": "generate_offender_story_ai_profiling_story__accused_id__get", "parameters": [{"name": "accused_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Accused Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StandardResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai/v2/ask": {"post": {"tags": ["ai-engine-v2"], "summary": "V2 Multi-Agent Investigation Query", "description": "20-agent investigation intelligence pipeline. Returns structured InvestigationResponse with evidence bundle, key findings, timeline, entity relationships, citations, chart specifications, and investigation recommendations. LLM is invoked only after evidence verification gate.", "operationId": "ask_v2_ai_v2_ask_post", "parameters": [{"name": "authorization", "in": "header", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Authorization"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AskRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/InvestigationResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai/v2/registry": {"get": {"tags": ["ai-engine-v2"], "summary": "List registered tool capabilities", "description": "Returns all tools registered in the ToolCapabilityRegistry with their descriptors.", "operationId": "list_tools_ai_v2_registry_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/ai/v2/health": {"get": {"tags": ["ai-engine-v2"], "summary": "V2 engine health check", "operationId": "health_v2_ai_v2_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": {"type": "string"}, "type": "object", "title": "Response Root  Get"}}}}}}}, "/health": {"get": {"summary": "Health", "operationId": "health_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": {"type": "string"}, "type": "object", "title": "Response Health Health Get"}}}}}}}, "/adapter-test": {"get": {"summary": "Test Adapter", "operationId": "test_adapter_adapter_test_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Response Test Adapter Adapter Test Get"}}}}}}}}, "components": {"schemas": {"AskRequest": {"properties": {"session_id": {"type": "string", "maxLength": 128, "minLength": 1, "title": "Session Id"}, "user_id": {"type": "string", "maxLength": 128, "minLength": 1, "title": "User Id"}, "question": {"type": "string", "maxLength": 5000, "minLength": 1, "title": "Question"}}, "type": "object", "required": ["session_id", "user_id", "question"], "title": "AskRequest"}, "Body_transcribe_audio_ai_transcribe_post": {"properties": {"audio": {"type": "string", "format": "binary", "title": "Audio"}}, "type": "object", "required": ["audio"], "title": "Body_transcribe_audio_ai_transcribe_post"}, "Body_upload_documents_ai_documents_upload_post": {"properties": {"file": {"type": "string", "format": "binary", "title": "File"}, "fir_id": {"type": "string", "title": "Fir Id"}}, "type": "object", "required": ["file"], "title": "Body_upload_documents_ai_documents_upload_post"}, "ChartSpec": {"properties": {"chart_id": {"type": "string", "title": "Chart Id"}, "chart_type": {"$ref": "#/components/schemas/ChartType"}, "title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}, "data": {"items": {"additionalProperties": true, "type": "object"}, "type": "array", "title": "Data"}, "x_axis": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "X Axis"}, "y_axis": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Y Axis"}, "color_by": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Color By"}, "config": {"additionalProperties": true, "type": "object", "title": "Config"}, "rationale": {"type": "string", "title": "Rationale", "default": ""}}, "type": "object", "required": ["chart_id", "chart_type", "title", "description"], "title": "ChartSpec", "description": "Frontend-consumable chart specification.\nVisualizationAgent generates specs, not images.\nThe React frontend renders these using Recharts / D3 / Leaflet."}, "ChartType": {"type": "string", "enum": ["timeline", "bar", "pie", "scatter", "network", "heatmap", "geo", "tree", "sankey", "line"], "title": "ChartType"}, "Citation": {"properties": {"source": {"type": "string", "title": "Source"}, "reference_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Reference Id"}, "snippet": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Snippet"}, "score": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Score"}}, "type": "object", "required": ["source"], "title": "Citation"}, "ErrorDetail": {"properties": {"code": {"type": "string", "title": "Code"}, "message": {"type": "string", "title": "Message"}, "details": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "Details"}}, "type": "object", "required": ["code", "message"], "title": "ErrorDetail"}, "EvidenceBundle": {"properties": {"ranked_items": {"items": {"$ref": "#/components/schemas/RankedEvidenceItem"}, "type": "array", "title": "Ranked Items"}, "top_evidence_text": {"type": "string", "title": "Top Evidence Text", "default": ""}, "overall_confidence": {"type": "number", "title": "Overall Confidence", "default": 0.0}, "confidence_label": {"type": "string", "title": "Confidence Label", "default": "none"}, "threshold_met": {"type": "boolean", "title": "Threshold Met", "default": false}, "evidence_count": {"type": "integer", "title": "Evidence Count", "default": 0}, "source_diversity": {"type": "integer", "title": "Source Diversity", "default": 0}}, "type": "object", "title": "EvidenceBundle", "description": "Output of EvidenceRankingAgent.\ntop_evidence_text is what gets sent to LLMReasoningAgent."}, "EvidenceScore": {"properties": {"confidence": {"type": "number", "title": "Confidence", "default": 0.0}, "freshness": {"type": "number", "title": "Freshness", "default": 1.0}, "source_reliability": {"type": "number", "title": "Source Reliability", "default": 0.5}, "relevance": {"type": "number", "title": "Relevance", "default": 0.5}, "composite": {"type": "number", "title": "Composite", "default": 0.0}}, "type": "object", "title": "EvidenceScore", "description": "Composite scoring for a single evidence item."}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "InvestigationResponse": {"properties": {"response_id": {"type": "string", "title": "Response Id"}, "session_id": {"type": "string", "title": "Session Id"}, "user_id": {"type": "string", "title": "User Id"}, "intent": {"type": "string", "title": "Intent"}, "complexity": {"type": "string", "title": "Complexity", "default": "simple"}, "executive_summary": {"type": "string", "title": "Executive Summary", "default": ""}, "key_findings": {"items": {"$ref": "#/components/schemas/KeyFinding"}, "type": "array", "title": "Key Findings"}, "timeline": {"items": {"$ref": "#/components/schemas/TimelineEvent"}, "type": "array", "title": "Timeline"}, "related_entities": {"items": {"$ref": "#/components/schemas/RelatedEntity"}, "type": "array", "title": "Related Entities"}, "evidence_bundle": {"anyOf": [{"$ref": "#/components/schemas/EvidenceBundle"}, {"type": "null"}]}, "citations": {"items": {"$ref": "#/components/schemas/Citation"}, "type": "array", "title": "Citations"}, "recommendations": {"items": {"type": "string"}, "type": "array", "title": "Recommendations"}, "chart_specs": {"items": {"$ref": "#/components/schemas/ChartSpec"}, "type": "array", "title": "Chart Specs"}, "confidence": {"type": "number", "title": "Confidence", "default": 0.0}, "confidence_label": {"type": "string", "title": "Confidence Label", "default": "none"}, "clarification_needed": {"type": "boolean", "title": "Clarification Needed", "default": false}, "clarification_question": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Clarification Question"}, "critic_passed": {"type": "boolean", "title": "Critic Passed", "default": true}, "critic_warnings": {"items": {"type": "string"}, "type": "array", "title": "Critic Warnings"}, "metadata": {"additionalProperties": true, "type": "object", "title": "Metadata"}, "v2": {"type": "boolean", "title": "V2", "default": true}}, "type": "object", "required": ["session_id", "user_id", "intent"], "title": "InvestigationResponse", "description": "Full V2 response schema returned by POST /ai/v2/ask.\nReplaces V1 StandardResponse for the V2 pipeline."}, "KeyFinding": {"properties": {"finding": {"type": "string", "title": "Finding"}, "evidence_ids": {"items": {"type": "string"}, "type": "array", "title": "Evidence Ids"}, "confidence": {"type": "number", "title": "Confidence", "default": 0.0}, "supported": {"type": "boolean", "title": "Supported", "default": true}}, "type": "object", "required": ["finding"], "title": "KeyFinding"}, "RankedEvidenceItem": {"properties": {"citation": {"$ref": "#/components/schemas/Citation"}, "scores": {"$ref": "#/components/schemas/EvidenceScore"}, "rank": {"type": "integer", "title": "Rank"}, "source_tier": {"type": "string", "title": "Source Tier", "default": "secondary"}}, "type": "object", "required": ["citation", "scores", "rank"], "title": "RankedEvidenceItem"}, "RelatedEntity": {"properties": {"entity_type": {"type": "string", "title": "Entity Type"}, "name": {"type": "string", "title": "Name"}, "role": {"type": "string", "title": "Role"}, "relationship_to_case": {"type": "string", "title": "Relationship To Case"}, "confidence": {"type": "number", "title": "Confidence", "default": 0.0}}, "type": "object", "required": ["entity_type", "name", "role", "relationship_to_case"], "title": "RelatedEntity"}, "ResponseMetadata": {"properties": {"intent": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Intent"}, "query_plan": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Query Plan"}, "evidence_sources": {"type": "integer", "title": "Evidence Sources", "default": 0}, "api_records": {"type": "integer", "title": "Api Records", "default": 0}, "langgraph_enabled": {"type": "boolean", "title": "Langgraph Enabled", "default": false}, "correlation_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Correlation Id"}, "confidence": {"type": "string", "title": "Confidence", "default": "low"}, "evidence_threshold_met": {"type": "boolean", "title": "Evidence Threshold Met", "default": false}, "evidence_required": {"type": "integer", "title": "Evidence Required", "default": 0}, "rag_citations": {"type": "integer", "title": "Rag Citations", "default": 0}, "sql_citations": {"type": "integer", "title": "Sql Citations", "default": 0}, "evidence_source_breakdown": {"additionalProperties": {"type": "integer"}, "type": "object", "title": "Evidence Source Breakdown"}, "persistence_enabled": {"type": "boolean", "title": "Persistence Enabled", "default": false}, "conversation_store_path": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Conversation Store Path"}}, "type": "object", "title": "ResponseMetadata"}, "StandardResponse": {"properties": {"success": {"type": "boolean", "title": "Success"}, "message": {"type": "string", "title": "Message", "default": ""}, "data": {"additionalProperties": true, "type": "object", "title": "Data"}, "metadata": {"$ref": "#/components/schemas/ResponseMetadata"}, "citations": {"items": {"$ref": "#/components/schemas/Citation"}, "type": "array", "title": "Citations"}, "errors": {"anyOf": [{"items": {"$ref": "#/components/schemas/ErrorDetail"}, "type": "array"}, {"type": "null"}], "title": "Errors"}}, "type": "object", "required": ["success"], "title": "StandardResponse"}, "TimelineEvent": {"properties": {"timestamp": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Timestamp"}, "event": {"type": "string", "title": "Event"}, "source": {"type": "string", "title": "Source"}, "confidence": {"type": "number", "title": "Confidence", "default": 0.0}, "citation_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Citation Id"}}, "type": "object", "required": ["event", "source"], "title": "TimelineEvent"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}